==> 02_ETL_australian_users_items.ipynb <==
import pandas as pd
import numpy as np
import json

# Se cargan los archivos originales
nombre_archivo = 'Orig/PI_ML_OPS-FT/Recursos/australian_users_items.json'

# Se abre el archivo en modo lectura ('r') con codificación utf-8
with open(nombre_archivo, 'r', encoding='utf-8') as archivo:
    # Se lee todas las líneas del archivo y almacenarlas en la lista 'datos'
    datos = archivo.readlines()

# Se convierte las líneas en registros utilizando eval y eliminando los espacios alrededor
registros = [eval(line.strip()) for line in datos]

# Se crea un DataFrame de pandas a partir de los registros
df_UserItems = pd.DataFrame(registros)

df_UserItems

# Vemos la informacion general del dataframe, indica que no hay valores nulos. 
df_UserItems.info()

# Vemos que la columna 'items' esta anidada
df_UserItems['items'][0]

# Normalizamos la columna 'items'
df_UserItems = pd.json_normalize(registros, record_path=['items'], meta=['user_id', 'items_count','steam_id','user_url'])

df_UserItems

# Se eliminan las columnas que no serán tenidas en cuenta en el modelamiento

df_UserItems.drop(['item_name', 'user_url'],  axis=1, inplace=True)

# Se verifica si hay registros duplicados y se eliminan
df_UserItems = df_UserItems.drop_duplicates()
# Se eliminan los registros totalmente nulos
df_UserItems= df_UserItems.dropna(how='all')

#Se borran los registros de la columna playtime_forever que tengan 0 
df_UserItems = df_UserItems.drop(df_UserItems[df_UserItems['playtime_forever'] == 0].index)

# Se convierte la columna 'playtime_forever' de minutos a horas
df_UserItems['playtime_forever']=df_UserItems['playtime_forever']/60

df_UserItems

# Se redondean los valores a dos decimales
df_UserItems['playtime_forever'] = df_UserItems['playtime_forever'].round(2)

# Se elimina la columna 'playtime_2weeks' y 'steam_id' por no ser relevante
df_UserItems = df_UserItems.drop(['playtime_2weeks', 'steam_id'], axis=1)
df_UserItems.head()

# Se guarda el archivo con formato csv 
df_UserItems.to_csv('user_items_cleaned.csv', index=False)
