==> 03_ETL_output_steam_games.ipynb <==
import pandas as pd
import numpy as np
import json

# Se cargan los archivos originales
nombre_archivo = 'Orig/PI_ML_OPS-FT/Recursos/output_steam_games.json'

# Se abre el archivo en modo lectura ('r') con codificación utf-8
with open(nombre_archivo, 'r', encoding='utf-8') as archivo:
    # Leer todas las líneas del archivo
    lineas = archivo.readlines()

# Se carga objetos JSON individualmente
objetos_json = [json.loads(line) for line in lineas]

# Se crea DataFrame
df_SteamGames = pd.DataFrame(objetos_json)

df_SteamGames

print('Informacion general del dataframe\n')
df_SteamGames.info()

# Se eliminan las filas con todos los valores nulos
df_SteamGames=df_SteamGames.dropna(how='all').reset_index(drop=True)
df_SteamGames.head()

# Se reemplazan valores vacíos, 'null' y 'None' con NaN en todo el DataFrame
df_SteamGames.replace(['', 'null', 'None'], np.nan, inplace=True)

# Se renombran columnas
df_SteamGames.rename(columns={'app_name': 'name', 'id': 'item_id'}, inplace=True)

# Se cuentan la cantidad de valores nulos en las columnas 'publisher' y 'developer'
print(df_SteamGames[['publisher', 'developer']].isnull().sum())

# Se cuentan la cantidad de valores nulos en las columnas 'name' y 'title'
print(df_SteamGames[['name', 'title']].isnull().sum())

valores_nulos_por_columna = df_SteamGames.isnull().sum()

# Muestra de los resultados
print(valores_nulos_por_columna)

# Se filtran registros donde item_id es nulo o NaN
df_SteamGames[df_SteamGames['item_id'].isna()]

# Se eliminan las filas donde item_id es nulo o NaN
df_SteamGames=df_SteamGames.dropna(subset=['item_id'])

# Se filtran registros donde 'name' es nulo o NaN
df_SteamGames[df_SteamGames['name'].isna()]

# Se elimina la columna 'title' ya que la columna 'name' contiene los mismos valores pero con 1 solo nulo. 
df_SteamGames.drop(['title'] ,axis=1, inplace=True)

# Se convierte 'release_date' a tipo datetime
df_SteamGames['release_date'] = pd.to_datetime(df_SteamGames['release_date'], errors='coerce', exact=False)

# Se crea la columna 'year' a partir de 'release_date'
df_SteamGames['year'] = df_SteamGames['release_date'].dt.year.astype('Int64')



# Se imprime información sobre nulos en 'year' después de la conversión
print("Nulos después de la conversión a datetime:")
print(df_SteamGames['year'].isnull().sum())

# Si aun hay valores nulos después de la interpolación, se llenan con la mediana. 
df_SteamGames['year'] = df_SteamGames['year'].fillna(df_SteamGames['year'].median())

# Volvemos a verificar si hay valores nulos en la columna 'year'
hay_nulos_en_year = df_SteamGames['year'].isnull().any()

if hay_nulos_en_year:
    print("Hay valores nulos en la columna 'year'.")
else:
    print("No hay valores nulos en la columna 'year'.")

#  Se asigna los valores de la columna 'publisher' a la columna 'developer' en las filas donde 'developer' es nula y 'publisher' no es nula.
df_SteamGames.loc[df_SteamGames['developer'].isnull() & ~df_SteamGames['publisher'].isnull(), 'developer'] = df_SteamGames['publisher']

# Se eliminan las siguientes columnas 'title','url','reviews_url','early_access','publisher','release_date' por considerarse irrelevantes
df_SteamGames.drop(['url','reviews_url','early_access','publisher','release_date'] ,axis=1, inplace=True)

df_SteamGames.head()

# Se convierte 'price' a tipo numérico, asignar 0.0 a 'Free To Play'
df_SteamGames['price'] = pd.to_numeric(df_SteamGames['price'], errors='coerce')
df_SteamGames['price'] = df_SteamGames['price'].fillna(0)

df_SteamGames.info()

genres = df_SteamGames['genres']
tags = df_SteamGames['tags']
specs = df_SteamGames['specs']

print('Genres:', genres)
print('Tags:', tags)
print('Specs:', specs)

# Se llenan los nulos de la columna genres con corchetes para poder transformar a listas
df_SteamGames['genres'] = df_SteamGames['genres'].fillna('[]')

# Se transforma la columna genero a listas, se la desanida y utilizamos get dummies para darle un valor numerico
genres_df = pd.DataFrame(df_SteamGames['genres'].tolist())
genres_df_obj = genres_df.stack()
genres_df1 = pd.get_dummies(genres_df_obj)
genres_df1 = genres_df1.groupby(level= [0], axis= 0).sum()

# Se evalua el aporte que tiene cada genero para analizar con cual quedarnos
genres_df1.sum().sort_values(ascending=False)/len(genres_df1)

# Se seleccionan las columnas que me voy a quedar
genres = genres_df1[['Indie','Action','Casual','Adventure','Strategy','Simulation','RPG','Free to Play','Early Access' ,'Sports','Massively Multiplayer','Racing'  ,'Design &amp; Illustration' ,'Utilities']]

# Se unen los dos dataframe y se elimina la columna de genres
df_SteamGames = pd.concat([df_SteamGames, genres],axis=1)
df_SteamGames.drop(columns=['genres'],inplace=True)

# Se eliminan las filas que tienen valores nulos en la columna 'Indie'.
df_SteamGames.dropna(subset='Indie',inplace=True)

# Se cambia la columna 'item_id' a tipo int
df_SteamGames['item_id'] = df_SteamGames['item_id'].astype('Int64')

df_SteamGames

# Se eliminará todas las filas en las que el valor de la columna item_id es nulo 
df_SteamGames.dropna(subset=['item_id'], inplace=True)

# Poniendo minúsculas los valores del campo 'developer' como estandarización
df_SteamGames['developer'] = df_SteamGames['developer'].str.lower()

# Se eliminan 'tags' y 'specs' para aligerar el modelo ya que no son necesarios
df_SteamGames.drop(columns=['tags', 'specs'],inplace=True)

df_SteamGames

df_SteamGames.info()

# Se guarda el archivo con formato csv
df_SteamGames.to_csv('steam_games_cleaned.csv', index=False)
