==> 05_Modelo_item.ipynb <==
import pandas as pd
from sklearn.metrics.pairwise        import cosine_similarity
from sklearn.metrics.pairwise        import linear_kernel
from sklearn.feature_extraction.text import TfidfVectorizer

df_SteamGames=pd.read_csv('steam_games_cleaned.csv')
df_UserReviews=pd.read_csv('user_reviews_cleaned.csv')

# Se crea un nuevo DataFrame llamado df_games que contiene solo las columnas relevantes
game_columns = ['item_id', 'name']
df_games = df_SteamGames[game_columns]

# Se hace un join entre df_games y df_UserItems utilizando 'item_id' como clave
df_joined = pd.merge(df_games, df_UserReviews[['item_id', 'review']], on='item_id', how='inner')

#Se crea una muestra para el modelo
muestra = df_joined.head(20000)

# Se crea el modelo de machine learning con Scikit-Learn
tfidf = TfidfVectorizer(stop_words='english')
muestra=muestra.fillna("")

tdfid_matrix = tfidf.fit_transform(muestra['review'])
cosine_similarity = linear_kernel( tdfid_matrix, tdfid_matrix)

### Función recomendacion_juego()
Ingresando el id de producto (juego) recibiremos una lista con 5 juegos recomendados similares

# Se crea la funcion de recomendación de 5 juegos recomendados para el juego ingresado.
def recomendacion_juego(id: int):
    if id not in muestra['item_id'].values:
        return {'mensaje': 'No existe el id del juego'}
    
    titulo = muestra.loc[muestra['item_id'] == id, 'name'].iloc[0]
    idx = muestra[muestra['name'] == titulo].index[0]
    sim_cosine = list(enumerate(cosine_similarity[idx]))
    sim_scores = sorted(sim_cosine, key=lambda x: x[1], reverse=True)
    sim_ind = [i for i, _ in sim_scores[1:6]]
    sim_juegos = muestra['name'].iloc[sim_ind].values.tolist()
    return {f'juegos recomendados para {id}': list(sim_juegos)}

# Ejemplo de uso del modelo 1 (2300 - DOOM II)
print(recomendacion_juego(2300))

# Ejemplo de uso del modelo 2 (21000 - LEGO Batman)
print(recomendacion_juego(21000))

muestra
