==> 07_CSV_to_PARQUET.ipynb <==
import pandas as pd
import pyarrow.parquet as pq

# Rutas del CSV
steam_games_cleaned_csv_path = "steam_games_cleaned.csv"
user_items_cleaned_csv_path = "user_items_cleaned.csv"
user_reviews_cleaned_csv_path = "user_reviews_cleaned.csv"

# CSV -> Parquet
df_steam_games_cleaned = pd.read_csv(steam_games_cleaned_csv_path)
df_user_items_cleaned = pd.read_csv(user_items_cleaned_csv_path)
df_user_reviews_cleaned = pd.read_csv(user_reviews_cleaned_csv_path)

df_steam_games_cleaned.to_parquet("df_steam_games_cleaned.parquet", index=False)
df_user_items_cleaned.to_parquet("df_user_items_cleaned.parquet", index=False)
df_user_reviews_cleaned.to_parquet("df_user_reviews_cleaned.parquet", index=False)

# Ruta del PARQUET
archivo_parquet = "df_steam_games_cleaned.parquet"

# Carga el archivo Parquet
tabla_parquet = pq.read_table(archivo_parquet)

# Muestra las columnas de la tabla
tabla_parquet
